// src/interfaces/global.ts
export interface GlobalModelIF {
  dispatch: Function;
  fold: boolean | null;
  explainShow: boolean;
  test: any;
  currentNumber: number;
}

export interface EffectsPCEIF {
  put: Function;
  call: Function;
  select: Function;
}

export interface ResponseGenerator {
  config?: any;
  data?: any;
  headers?: any;
  request?: any;
  status?: number;
  statusText?: string;
}